Separate hydrogen and internal components
